get rid of extra file
