+++ packages/console/app/src/routes/auth/logout.ts
@@ -1,6 +1,6 @@
 import { redirect } from "@solidjs/router"
 import { APIEvent } from "@solidjs/start"
-import { useAuthSession } from "~/context/auth.session"
+import { useAuthSession } from "~/context/auth"
 
 export async function GET(event: APIEvent) {
   const auth = await useAuthSession()
