+++ packages/console/app/src/routes/auth/status.ts
@@ -1,5 +1,5 @@
 import { APIEvent } from "@solidjs/start"
-import { useAuthSession } from "~/context/auth.session"
+import { useAuthSession } from "~/context/auth"
 
 export async function GET(input: APIEvent) {
   const session = await useAuthSession()
