+++ sst-env.d.ts
@@ -156,6 +156,10 @@ declare module "sst" {
       "type": "sst.sst.Secret"
       "value": string
     }
+    "ZEN_SESSION_SECRET": {
+      "type": "sst.sst.Secret"
+      "value": string
+    }
     "ZenData": {
       "name": string
       "type": "sst.cloudflare.Bucket"
